Shop_site/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-filter==24.3
Pillow==10.4.0
django-cors-headers==4.3.1
requests==2.32.3
